refactor: modularize ui and session logic
